fix: Check the entier vmid when getting VM
The code was assuming the first VM matched (or that only one is
returned) when searching by the custom vmid field.  Netbox
actually does a partial match, so searching for a longer ID
would also return matching shorter ones.

This change will iterate through the returned VMs looking for an
exact match.