feat: Add the ability to sync with Proxmox
Adds a new ProxmoxProvider

Configuration is now modified to accept an environment variable of
PROVIDER.  If PROVIDER=Proxmox the proxmox provider will be used,
otherwise the VmWareProvider is used.

Proxmox does not provide the IP addresses.
The Proxmox provider also requires the use of an API token.

feat: Add IPs to proxmox VMs

If the agent is installed this will allow the IPs to be set.  It
will only work on VMs.  LXC containers will not report IP.